refactor: 활동 요일 ENUM을 PARTY DOMAIN 아래로 변경
