+++ src/main/java/umc/cockple/demo/domain/party/dto/PartyCreateDTO.java
@@ -8,14 +8,13 @@ import jakarta.validation.constraints.Size;
 import lombok.Builder;
 import umc.cockple.demo.domain.party.exception.PartyErrorCode;
 import umc.cockple.demo.domain.party.exception.PartyException;
-import umc.cockple.demo.global.enums.ActiveDay;
+import umc.cockple.demo.domain.party.enums.ActiveDay;
 import umc.cockple.demo.global.enums.ActivityTime;
 import umc.cockple.demo.global.enums.Level;
 import umc.cockple.demo.global.enums.ParticipationType;
 import umc.cockple.demo.global.exception.GeneralException;
 
 import java.time.LocalDateTime;
-import java.util.ArrayList;
 import java.util.List;
 
 public class PartyCreateDTO {
